% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{with_mock}
\alias{with_mock}
\title{Mock functions in a package.}
\usage{
with_mock(..., mock_env = pkg_env(), eval_env = parent.frame())
}
\arguments{
\item{...}{Named parameters redefine mocked functions, unnamed parameters
will be evaluated after mocking the functions}

\item{mock_env}{The environment in which to patch the functions,
defaults to the top-level environment. A string is interpreted as
package name.}

\item{eval_env}{Environment in which expressions passed as \code{...} are
evaluated, defaults to \code{\link[=parent.frame]{parent.frame()}}.}
}
\value{
The result of the last unnamed argument passed as \code{...} (evaluated
in the environment passed as \code{eval_env})
}
\description{
Mocking allows you to temporary replace the implementation of functions
within a package, which useful for testing code that relies on functions
that are slow, have unintended side effects or access resources that may
not be available when testing.

Up until recently, such capability was offered via \code{\link[testthat:with_mock]{testthat::with_mock()}},
but with release of version 3.0.0 and introduction of edition 3, this was
deprecated from testthat, leaving it to third party packages to replace this
feature. This mocking implementation is powered by
\code{\link[utils:getFromNamespace]{utils::assignInNamespace()}} and therefore, caveats outlined in the
corresponding documentation apply here too.
}
\examples{

url <- "https://eu.httpbin.org/get?foo=123"
mok <- function(...) "mocked request"

with_mock(
  `curl::curl_fetch_memory` = mok,
  curl::curl_fetch_memory(url)
)

dl_fun <- function(x) curl::curl_fetch_memory(x)

with_mock(
  `curl::curl_fetch_memory` = mok,
  dl_fun(url)
)

json <- function(...) '["mocked request"]'

with_mock(
  `curl::curl` = json,
  jsonlite::fromJSON(url)
)

with_mock(
  `curl::curl` = json,
  parse_and_simplify = function(txt, ...) gsub('\\\\[?\\\\"\\\\]?', "", txt),
  jsonlite::fromJSON(url),
  mock_env = "jsonlite"
)

}
