% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{with_mock}
\alias{with_mock}
\title{Mock functions in a package.}
\usage{
with_mock(..., mock_env = pkg_env(), eval_env = parent.frame())
}
\arguments{
\item{...}{Named parameters redefine mocked functions, unnamed parameters
will be evaluated after mocking the functions}

\item{mock_env}{The environment in which to patch the functions,
defaults to the top-level environment.  A character is interpreted as
package name.}

\item{eval_env}{Environment in which expressions passed as \code{...} are
evaluated, defaults to \code{\link[=parent.frame]{parent.frame()}}.}
}
\value{
The result of the last unnamed argument passed as \code{...} (evaluated
in the environment passed as \code{eval_env})
}
\description{
Mocking allows you to temporary replace the implementation of functions
within a package, which useful for testing code that relies on functions
that are slow, have unintended side effects or access resources that may
not be available when testing.

This works by using some C code to temporarily modify the mocked function
\emph{in place}. On exit, all functions are restored to their previous state.
This is somewhat abusive of R's internals so use with care. In particular,
functions in base packages cannot be mocked; to work around you'll need to
make a wrapper function in your own package..
}
\examples{

url <- "https://eu.httpbin.org/get?foo=123"
mok <- function(...) "mocked request"

with_mock(
  `curl::curl_fetch_memory` = mok,
  curl::curl_fetch_memory(url)
)

dl_fun <- function(x) curl::curl_fetch_memory(x)

with_mock(
  `curl::curl_fetch_memory` = mok,
  dl_fun(url)
)

json <- function(...) '["mocked request"]'

with_mock(
  `curl::curl` = json,
  jsonlite::fromJSON(url)
)

with_mock(
  `curl::curl` = json,
  parse_and_simplify = function(txt, ...) gsub('\\\\[?\\\\"\\\\]?', "", txt),
  jsonlite::fromJSON(url),
  mock_env = "jsonlite"
)

}
